added provisioning

+++ adal-auth-ubuntu/aad-login/install.sh
@@ -44,7 +44,10 @@ apt-get install npm
 npm install
 apt-get install nodejs
 apt-get --purge remove node
-ln -s /usr/bin/nodejs /usr/bin/node
+if [ ! -f /etc/pam.d/common-auth ] 
+then
+  ln -s /usr/bin/nodejs /usr/bin/node
+fi
 
 sed -i 's/aadDirName/'$aadDirName'/' aad-login.js
 sed -i 's/aadClientID/'$aadClientID'/' aad-login.js
